Add initial app styling
