+++ src/PersonalWebSpace_frontend/components/Footer.svelte
@@ -1,0 +1,3 @@
+<footer class='py-7 space-y-6 items-center text-center' id='about'>
+  <p>Powered by <a href='https://vdfyi-uaaaa-aaaai-acptq-cai.ic0.app/' target='_blank' rel="noreferrer" class='underline'>Open Internet Metaverse</a> and hosted on <a href='https://internetcomputer.org/' target='_blank' rel="noreferrer" class='underline'>Internet Computer</a></p>
+</footer>
