+++ src/PersonalWebSpace_frontend/components/Topnav.svelte
@@ -1,0 +1,6 @@
+<div class="flex items-center justify-around flex-wrap bg-slate-600 p-3 text-l font-bold text-white">
+  <a href='#top'>Open Internet Metaverse</a>
+  <a href='#create'>Create</a>
+  <a href='#spaces'>My Spaces</a>
+  <a href='#/explore'>Explore</a>
+</div>
